mostly done
I got everything initialized I think and fixed a few issues with the naming of some of variables. I added HazyEndArm which is the code for the ending arm. All you have to do for that one is add in the code for how long the fold command needs to move in order to work. We will need to make it work by only moving the motor when its being pressed. I started to set up the OI and the buttons, but we need to talk to drive team to see how they went each controller set up. We also still need to create the code for the shooter, which has not been put in here yet. Idk, there might be a few problems with the current code, but we will see